#6581: split up sender/receiver
