+++ src/contentScript/pageEditor.ts
@@ -24,7 +24,7 @@ import {
 import { type BrickConfig } from "@/bricks/types";
 import { cloneDeep } from "lodash";
 import ConsoleLogger from "@/utils/ConsoleLogger";
-import { type SerializableResponse } from "@/pageScript/messenger/pigeon";
+import { type SerializableResponse } from "@/types/messengerTypes";
 import apiVersionOptions from "@/runtime/apiVersionOptions";
 import { clearDynamicElements } from "@/contentScript/pageEditor/dynamic";
 import { reactivateTab } from "./lifecycle";
