+++ src/pageScript/messenger/sender.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 PixieBrix, Inc.
+ * Copyright (C) 2023 PixieBrix, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -15,20 +15,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { serializeError } from "serialize-error";
-import {
-  type ActionType,
-  type Message,
-  type Meta,
-  type SerializedError,
-} from "@/types/messengerTypes";
-import { cleanValue } from "@/utils/objectUtils";
 import { once } from "lodash";
 import { logPromiseDuration } from "@/utils/promiseUtils";
 import injectScriptTag from "@/utils/injectScriptTag";
 import { expectContext } from "@/utils/expectContext";
-
-/** @file The first messenger before webext-messenger. Deprecated, see https://github.com/pixiebrix/webext-messenger/issues/5 */
 
 type SendScriptMessage<TReturn = unknown, TPayload = unknown> = (
   payload: TPayload
@@ -53,6 +43,7 @@ function assertCustomEvent(
 
 const injectPageScriptOnce = once(async (): Promise<void> => {
   expectContext("contentScript");
+
   console.debug("Injecting page script");
   const script = await logPromiseDuration(
     "injectPageScript",
@@ -85,8 +76,8 @@ export function createSendScriptMessage<TReturn = unknown, TPayload = unknown>(
       assertCustomEvent(type, event);
 
       const { id } = event.detail;
-      // This listener also receives it's own FULFILLED/REJECTED messages it sends back to the content
-      // script. So if you add any logging outside of the if, the logs are confusing.
+      // This listener also receives its own FULFILLED/REJECTED messages it sends back to the content
+      // script. So if you add any logging outside the `if`, the logs are confusing.
       const callback = callbacks.get(id);
 
       if (callback != null) {
@@ -133,122 +124,5 @@ export function createSendScriptMessage<TReturn = unknown, TPayload = unknown>(
     );
 
     return promise as Promise<TReturn>;
-  };
-}
-
-// eslint-disable-next-line @typescript-eslint/ban-types -- Line can be dropped once we migrate to `webext-messenger`
-export type SerializableResponse = boolean | string | number | object | void;
-
-export interface RemoteProcedureCallRequest<TMeta extends Meta = Meta>
-  extends Message<ActionType, TMeta> {
-  payload: unknown[];
-}
-
-export type HandlerOptions = {
-  asyncResponse?: boolean;
-};
-
-type Handler = (...args: unknown[]) => Promise<SerializableResponse>;
-
-export type HandlerEntry = {
-  handler: Handler;
-  options: HandlerOptions;
-};
-
-interface ErrorResponse {
-  $$error: SerializedError;
-}
-
-export function isNotification({
-  asyncResponse = true,
-}: HandlerOptions = {}): boolean {
-  return !asyncResponse;
-}
-
-export function isErrorResponse(ex: unknown): ex is ErrorResponse {
-  return typeof ex === "object" && ex != null && "$$error" in ex;
-}
-
-export function toErrorResponse(
-  requestType: string,
-  ex: unknown
-): ErrorResponse {
-  return { $$error: serializeError(ex) };
-}
-
-export function isRemoteProcedureCallRequest(
-  message: unknown
-): message is RemoteProcedureCallRequest {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any -- This is a type guard function and it uses ?.
-  return typeof (message as any)?.type === "string";
-}
-
-type AttachHandler = (type: string, handler: Handler) => void;
-
-// Message Type -> Handler
-const handlers = new Map<string, Handler>();
-
-async function messageHandler(event: MessageEvent): Promise<void> {
-  const type: string = event.data?.type;
-
-  const handler = handlers.get(type);
-
-  if (handler == null) {
-    return;
-  }
-
-  const { meta, payload } = event.data;
-
-  console.debug("RECEIVE", type, event.data);
-
-  try {
-    const result = await handler(payload);
-
-    let cleanResult;
-    try {
-      // Chrome will drop the whole detail if it contains non-serializable values, e.g., methods
-      cleanResult = cleanValue(result ?? null);
-    } catch (error) {
-      console.error("Cannot serialize result", { result, error });
-      throw new Error(`Cannot serialize result for result ${type}`);
-    }
-
-    const detail = {
-      id: meta.id,
-      result: cleanResult,
-    };
-    console.debug("pageScript responding %s_FULFILLED", type, detail);
-    document.dispatchEvent(
-      new CustomEvent(`${type}_FULFILLED`, {
-        detail,
-      })
-    );
-  } catch (error) {
-    try {
-      const detail = {
-        id: meta.id,
-        error,
-      };
-      console.warn("pageScript responding %s_REJECTED", type, detail);
-      document.dispatchEvent(
-        new CustomEvent(`${type}_REJECTED`, {
-          detail,
-        })
-      );
-    } catch (error_) {
-      console.error(
-        "An error occurred while dispatching an error for %s",
-        type,
-        { error: error_, originalError: error }
-      );
-    }
-  }
-}
-
-export function initialize(): AttachHandler {
-  window.addEventListener("message", messageHandler);
-
-  return (messageType: string, handler: Handler) => {
-    handlers.set(messageType, handler);
   };
 }
