+++ src/tsconfig.strictNullChecks.json
@@ -267,7 +267,8 @@
     "./pageScript/frameworks/errors.ts",
     "./pageScript/frameworks/scanner.ts",
     "./pageScript/messenger/constants.ts",
-    "./pageScript/messenger/pigeon.ts",
+    "./pageScript/messenger/sender.ts",
+    "./pageScript/messenger/receiver.ts",
     "./permissions/patterns.ts",
     "./permissions/permissionsTypes.ts",
     "./runtime/apiVersionOptions.ts",
@@ -376,6 +377,7 @@
     "./utils/injectIframe.tsx",
     "./utils/injectScriptTag.tsx",
     "./utils/injectStylesheet.tsx",
+    "./utils/legacyMessengerUtils.ts",
     "./utils/notify.tsx",
     "./utils/objToYaml.ts",
     "./utils/objectUtils.ts",
